notifications are split into redux and overmind versions
